auth blueprint theoretically finished
