+++ myproject/app/auth/templates/register.html
@@ -17,7 +17,7 @@
 <div class="largediv">
 
     <h2 class="text-center mb-3">Sign up</h2>
-    <form action="{{ url_for('register_handle') }}" method="post"
+    <form action="{{ url_for('auth.register_handle') }}" method="post"
           class="w-50 gap-3" enctype="multipart/form-data">
         {{ form.csrf_token() }}
